improve the formatting of parenthesized expressions
